Added some printing and FeeOption class.
Added some printing because there was some hinky stuff going on with transfering the whole name. @dombo helped me figure it out and it ended up being that the options weren't being sent at all. So editted the RpcService to do that as well. Keeping the printing in.

I'll probably make a "Paramaters" Namespace Just because it seems that some of the functions need objects not strings. So better to add it to the library later.